backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.31.0
wandb==0.15.4
psycopg2-binary==2.9.6
